requirements.txt
Flask==3.1.2
gunicorn==21.2.0
google-generativeai==0.3.0
